backend/requirements.txt
fastapi
uvicorn
python-multipart
# mediapipe
# tensorflow
firebase-admin
pydantic-settings
opencv-python
joblib


